Move 'make_equilibrium' into EUDEMO equilibria
